Introduce better connect behavior
This commit causes several changes to the internals of this
library. We are now using the python builtin modules "sched"
and "queue" to send and receive data. The connect function
is now blocking until we reach the connect_timeout or the
server sends the end of the motd.

There are two new configuration variables:

* connect_timeout
* reconnect_interval